requirements.txt
einops
huggingface-hub
imageio
matplotlib
ninja
opencv-python
opencv-contrib-python
pandas
safetensors
scipy
timm
tqdm
wandb
neptune
scikit-image
plyfile
torchmetrics
omegaconf
hydra-core
jaxtyping
hydra-submitit-launcher
submitit
GitPython
lightning
pytubefix
ipdb